requirements.txt
httpx
polars
duckdb
pyyaml
python-dotenv
tenacity
